🚧 PFXN5E task: Split branch_pr list-state test
